requirements.txt
google-generativeai==0.3.2
Pillow==10.2.0
python-dotenv==1.0.1
streamlit==1.32.0
requests==2.31.0
numpy==1.26.4
pandas==2.2.1
python-magic==0.4.27
python-slugify==8.0.4
watchdog==4.0.0 